Fix animateptr and interpolation 64-bit pointer corruption
On 64-bit Linux, long is 8 bytes but sector/wall struct fields
(ceilingz, floorz, x, y) are int32_t (4 bytes). Dereferencing
long* pointers to these fields reads/writes 8 bytes from 4-byte
fields, corrupting adjacent struct members.

Changes:
- animateptr: long* -> int32_t* (points at int32_t struct fields)
- animategoal, animatevel: long -> int32_t (hold field values)
- curipos: long* -> int32_t* (interpolation pointers to fields)
- oldipos, bakipos: long -> int32_t (hold interpolated values)
- setinterpolation/stopinterpolation: long* -> int32_t* params
- getanimationgoal/setanimation: long* -> int32_t* params
- MENUES.C save/load: use sizeof(animateptr[0]) and intptr_t
  for correct pointer-offset arithmetic on 64-bit

+++ source/ACTORS.C
@@ -39,7 +39,7 @@ void updateinterpolations()  //Stick at beginning of domovethings
 }
 
 
-void setinterpolation(long *posptr)
+void setinterpolation(int32_t *posptr)
 {
 	long i;
 
@@ -51,7 +51,7 @@ void setinterpolation(long *posptr)
 	numinterpolations++;
 }
 
-void stopinterpolation(long *posptr)
+void stopinterpolation(int32_t *posptr)
 {
 	long i;
